chore: rename plan file to fix typo
