+++ src/context/plan.js
@@ -1,7 +1,7 @@
 const processPlan = require('./process-plan');
 const constantsPlan = require('./constants-plan');
 const envPlan = require('./env-plan');
-const dependenciesPlan = require('./dependancies-plan');
+const dependenciesPlan = require('./dependencies-plan');
 const utilsPlan = require('./utils-plan');
 const serializersPlan = require('./serializers-plan');
 const servicesPlan = require('./services-plan');
